都道府県一覧とcheckboxを生成するコンポーネントを作成
