+++ src/components/CheckList.jsx
@@ -1,0 +1,19 @@
+
+
+const style = {
+    display: "inline-block",
+    width: "100px",
+  }
+
+export const CheckList = ({ prefectures, onChange }) => {
+
+    return (
+        <div >
+        {
+            prefectures.map(post => <li style={style} key={post.prefCode}>
+            <input type="checkbox" name={post.prefName} onChange={onChange} id={"checkbox" + post.prefCode}></input> 
+            {post.prefName}</li>)
+        }
+      </div>
+    );
+}
